+++ packages/governance-core/src/index.ts
@@ -2,3 +2,4 @@ export * from "./types.js";
 export * from "./evaluateGovernedAction.js";
 export * from "./evaluateGovernedRuntimeAction.js";
 export * from "./evaluateGovernedRuntimeActionWithReceipt.js";
+export * from "./createAgencyFingerprintForGovernedRuntimeAction.js";